requirements.txt
facebook-sdk
python-twitter
requests